Reorder integration tests
